Better support for single-element lists
Our parser is a little schizophrenic.  If a list has a single element,
it's imported as [element], if it has multiple elements it's imported as
[[element, element]].  That extra list is there for no good reason, and
affects all atpkg vector types.

We fix the parser to import all lists as [element, element], and remove
some workarounds we had in place specific to source files, which did not
exactly work and also did not account for vector types that aren't
source files.